Add front/back connectivity and CORS support for development
Now the React frontend can call the Django backend to log users in and gather user data. On login, the user's Authentication Token is stored in sessionStorage. When calling Django for user data, the Authentication Token is attached to the request to allow the user to be identified. On logout, sessionStorage is cleared and the user is redirected to the login page.