Add handleToken prop so ProfileNav can change upper state and force re-render
Added handleToken prop containing the useState method for authenticationToken in ProfileApp. This prop was added to all parent components of ProfileNav to allow it to change state in ProfileApp, thereby forcing a re-render and redirecting the user to the login page through reach/router's Redirect component. This is a crude implemenation, should look to use Redux or some other global state management for user Authentication Tokens.