Fix crashing issue related to selectedReplayHash value
When logging in if a user navigated to Replays too quickly it would cause React to crash. React would also crash after a user uploaded a replay and navigated to the Replay section too quickly. Believe this has something to do with selectedReplayHash being stored in the Redux store whereas other replay information isn't, causing null values for other info but a non-null value for selectedReplayHash.