Full UI update and SideBar + Nav refactor
Updated the UI for the login page and the main app. Login page is now fully responsive as well. App UI was polished up and the ability to permanently extend the sidebar was added. Also added basic framework for choosing race data. The PageSidebar and PageNav components were refactored to allow the permanent sidebar extension to work. Needed to rewrite the hover state management with PageSidebar as the controlling component instead of having mixed state in PageSidebar and PageNav.