Refactor app to improve architecture
Restructured the app with the router now inside instead of as a wrapper. This means the nav section doesn't get re-rendered during navigation, which was messing up transitions. Also makes page components cleaner as they just export basic HTML now and are then slotted into the page structure in the PageTemplate component.