Replay Polling + Downloads + Saving Chart Selections (#3)
* Add replay download + save chart selections

Users can now download the selected replay. Chart selections for the replay timeline and the weekly trends are now saved in localStorage so selections persist on reload.

* Add polling for new replays

The backend is now polled every 30sec for the current number of replays for each race. If the returned value is different to the current value on the frontend, the replays for that race are refreshed.

* Add page icons + change nav styling

Added a unique icon for each page. Also made the styling for the nav less loud and more minimalistic

* Rename Analysis -> Trends and Icon -> NavIcon