Improve Trends Page (#5)
* Refactor FeatureVote to remove unnecessary code

Thanks to some advice from Reddit I removed a decent amount of code that either did nothing or was over-engineered. For instance: the disabled property on the submit button and tracking checkbox count.

* Disable button on Login + fix loading indicator bug

* Alter logout request to improve responsiveness

* Add more stats + 3mo period support + new line colour scheme

* Add abort controllers and errors for timeline requests

* Add winrate correlations feature

Shows which stats are most correlated with a player's winrate. Stats are currently measured from all games in the last 1yr.