Shareable Replay Pages (#6)
* Refactor colour system

* Misc style improvements

Changed the styling of: Download button to reduce use of primary blue and make it pop less, player red and blue to reduce vibration against the background color, header and replay list colours to break up the monotonous background color, replay record background in line with replay list colour change. Also added a user indicator.

* Add replay links and player info on timeline state + misc styling

Users can now share links that will show the replay on a page that is accessible to anyone. Also added basic player info to the timeline state area for ease of use. Added box shadow styling to player pills to distinguish the user.

* Refactor and split Replay component into Replay + ReplayView

MASSIVE refactor of the Replay component moving most of the subcomponents and presentational things into ReplayView to make it reusable for the shareable replay pages project.

* BIG refactor: custom hooks + restructure + cleanup

This is an absolutely massive refactor. The main architecture was rewritten to be hooks-focused, which makes dealing with data fetching lifecycles far easier and creates a more standard API to work with. Multiple components/folders have also been renamed and components have been moved for semantics. Work towards cleaning up the code in general has also been done such as enforcing prop destructuring and splitting large chunks of code into components.

* useFetch bug fix + useTrends hook + refactor actions

* Re-order useFetch params

* useFetch no longer requires tracking deps apparently

* Add Unspent Resources to timeline + fix minor styling issues

* Add supply count + army value + resources lost to timeline

* Change icons to official Blizzard art

* Improve timeline UI + add Burrowed and Flying states

Also fixed an issue with upgrades so that now only level upgrades show an indicator

* Fix minor Protoss upgrade issues + add Morphing unit state for Zerg units

* Add styling for Energy charts

* Remove metric names from Energy charts

* Add ability usage/targets and improve UI for RaceState

* Add tooltips to resource data in RaceState

* Add Creep data to RaceState

* Update StatCategory styling for replays + add new stats